Moving to Ubuntu 20.04 LTS
1. This is the first attempt to move to 20.04 (Focal Fossa).
Since in Ubuntu 20.04 LXD can installed via snap only and
is running in its own container, it brings a lot of redundancy
via binary dependencies to Bitscout, which inflates LiveCD image size.
That is why it was decided to move to a new lighter container management
based on systemd-nspawn.

2. Wicd network management tool package was not found in the repository,
so we might be switching to alternative WiFi manager soon.

3. Initrd file format has changed, which broke packing/unpacking routine. This will
be addressed soon.